refactor : context import 위치를 index로 변경
